final login, register, logout routes and logic

+++ controllers/coffeeController.js
@@ -15,8 +15,8 @@ const upload = multer({storage: storage})
 
 const getAllCoffees = async (req, res) => {
     try {
-        const coffees = await Coffee.find()
-        res.render('home', {coffees: coffees})
+        const coffees = await Coffee.find().populate('owner')
+        res.render('home', {coffees: coffees, user: req.user})
     } catch (error) {
         console.log(error)
     }
@@ -24,7 +24,7 @@ const getAllCoffees = async (req, res) => {
 
 
 const uploadPage = (req, res) => {
-    res.render('upload')
+    res.render('upload', {user: req.user})
 }
 
 const createCoffee = async (req, res) => {
@@ -32,7 +32,8 @@ const createCoffee = async (req, res) => {
         const coffee = new Coffee({
             name: req.body.name,
             blurb: req.body.blurb,
-            image: req.file.filename //multer places the file infor in the req.file
+            image: req.file.filename, //multer places the file infor in the req.file
+            owner: req.user._id
        })
 
        await coffee.save()
@@ -46,7 +47,7 @@ const createCoffee = async (req, res) => {
 const editPage = async (req, res) => {
     try {
         const coffee = await Coffee.findById(req.params.id)
-        res.render('edit', {coffee:coffee})
+        res.render('edit', {coffee:coffee, user: req.user})
     } catch (error) {
         console.log(error)
     }
@@ -54,7 +55,10 @@ const editPage = async (req, res) => {
 
 const updateCoffee = async (req, res) => {
     try {
-        await Coffee.findByIdAndUpdate(req.params.id, req.body)
+        let coffee = await Coffee.findById(req.params.id)
+        if(coffee.owner.equals(req.user._id)){
+            await Coffee.findByIdAndUpdate(req.params.id, req.body)
+        }
         res.redirect('/')
     } catch (error) {
         console.log(error)
@@ -63,7 +67,10 @@ const updateCoffee = async (req, res) => {
 
 const deleteCoffee = async (req, res) => {
     try {
-        await Coffee.findByIdAndRemove(req.params.id)
+        let coffee = await Coffee.findById(req.params.id)
+        if(coffee.owner.equals(req.user._id)){
+            await Coffee.findByIdAndRemove(req.params.id)
+        }
         res.redirect('/')
     } catch (error) {
         console.log(error)
